+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - intellij-latte-pro.xml
+- initFormData is optional
+- group label option
 
 
 ## [1.5.2] - 2023-01-20